Add "unrotate" parameter for each joint
This parameter is used to accomodate serial joints which do
not rotate with the previous joint.  This type of linkage is
common in some robot types (notably silicon wafer handlers).

In general, the parameter will be 0 (no compensation), 1 (add
prior joint position), or -1 (sibtract prior joint position).
It is possible to set it to any integer value, and other values may
make sense (if there is gearing, for example)