add patch to axis.py from a dropped merge
Commit 0662094ae97f32b2201b437e4c6d32e8f694bb3a merges master into
joints_axes3.  That merge commit (and all other merges) was dropped
during the rebase of joints_axes3 to joints_axes4.

That merge commit in ja3 also introduces the unrelated change contained
in this commit; this commit was added to ja4 (in the same place as
the 0662094ae merge, relative to the other commits in the branch) to
introduce the new patch introduced during the merge.