halui: fix some jogging bugs (adapt ja9_updates)
This fixes some bugs in halui, relating to jogging of the selected axis.
See SF bug #449.

Before this commit, when an axis is jogged using the
"halui.jog.selected.plus" pin (or the ".minus" pin), and then the selected
axis is changed (by sending a rising edge to a "halui.joint.N.select"
pin), then the old joint would keep jogging, and the new joint would
not start jogging.

In addition, if you later drive "halui.jog.selected.plus" (or ".minus")
false, the jogging joint would not stop jogging.

This commit fixes all of these bugs.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>