locking indexer:joint homing with gcode locking
Test with axis_9axis.ini (uses locking_indexer for axis B)

For joints_axes, a locking indexer must be a joint for homing
and must be unlockable for G0 gcode moves.  The joint number
to use for an axis cannot be predicted for arbitrary,
non-identity kinematics.

Imlementation:
 1) Use motmod unlock_joints_mask= parameter to create hal pins:
    joint.n.unlock, joint.n.is-locked
    Example:
       [EMCMOT]motmod unlock_joints_mask=0x10 (joint 4)
 3) Specify the joint to use for an axis with inifile item:
       [AXIS_L]LOCKING_INDEXER_JOINT=
    Example:
      [AXIS_A]LOCKING_INDEXER_JOINT = 4 (joint 4)

Note: There must be a one-to-one correspondence between the
      specified joint and the axis (A,B,C)

Also:
     clarify joint/axis notation in numerous function prototypes
     (dont refer to parameter as axis, use jnum etc.)

     Update configs/sim/axis/axis_9axis.ini

     WRAPPED_ROTARY,was: AXIS_3, is: AXIS_A ...

     axis.py: accomodate num_joints >= len(trajcoordinates)

Signed-off-by: Dewey Garrett <dgarrett@panix.com>