axis.py: Warn for duplicated coords (kins Inverse)
Kins which support duplicated coordinates (gantrykins and
gentrivkins kinstype=BOTH for instance) currently return
the value of the highest joint number for a duplicated
axis letter.  Moving a joint may not show as movement on
the display if there is a higher numbered joint for a
(duplicated) axis letter.

Alternatives inverse kinematics could be considered for these
'general-purpose' kinematics:

Example: compute the average of all joint values in use for a
         duplicated axis letter.

Signed-off-by: Dewey Garrett <dgarrett@panix.com>