inihal: new ini hal pins for home, home_offset  JA
also: new emcJointUpdateHomingParams()
      to update home & home_offset only

Thanks for the suggestion by pkmcnc (Andrew K)

Background:

An intrinsic problem of parallel robots is that there are
almost always some geometry errors in adapting the
mathematical model to a physical realization.  Moreover, when
using index homing, the exact location of the index may be
unknown.

Unlike a trivial kinematics machine, it is not always possible
to directly measure or compute exact values for [JOINT_*]HOME
& HOME_OFFSET.

A practical homing method for such cases is to:
 1) home to precalculated values for HOME & HOME_OFFSET
 2) manually move the platform to the correct home position
 3) calculate the difference in joint positions (with some hal
    components for example)
 4) update HOME_OFFSET (probably also HOME) to the corrected values.
 3) home again

Adjusting home position this way could be useful for some
identity kinematics machines as well.

Signed-off-by: Dewey Garrett <dgarrett@panix.com>