motion: move handling of EMCMOT_TELEOP to a function, for reuse
This is primarily a refactor to allow reuse of the switch-to-teleop code,
but it also slightly changes it:

Before this commit, the code would modify a bunch of state to indicate
the switch from free to teleop mode, *then* check if teleop was allowed
(by checking for unhomed joints and non-identity kins).  If teleop was
not allowed it would only partially restore the state it had modified.

This commit moves the check to the beginning, and doesn't modify any
state at all if Teleop mode is not allowed.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>